@@ -23,7 +23,9 @@ dependencies:
   connectivity_plus: ^6.0.3
   pocketbase: ^0.17.0
   stock: ^1.0.1
-  sentry_flutter: ^8.4.0
+  sentry: ^8.6.0
+  sentry_flutter: ^8.6.0
+  
   path_provider: ^2.1.3
   http: ^1.2.1
   html_unescape: ^2.0.0
@@ -36,7 +38,7 @@ dependencies:
   # cloud_firestore: ^4.14.0
   isolated_worker: ^0.1.1
   # adding realm to replace firestore on desktop in transporting data to cloud
-  realm: ^3.2.0
+  realm: ^3.3.0
   dart_mappable: ^4.2.1
   dio: ^5.4.3+1
   talker_dio_logger: ^4.1.5
@@ -56,9 +58,11 @@ dependency_overrides:
   meta: ^1.12.0
   analyzer: ^6.0.0
   
+  
 
 dev_dependencies:
   build_runner: ^2.4.8
   riverpod_lint: ^2.3.7
   dart_mappable_builder: ^4.2.1
   custom_lint: ^0.6.4
+  mockito: ^5.4.4
